Added initial values for all date form fields.
